made the schema.prisma file, and migrated the database. it should now be migrated correctly. also started to seed database process, added in topics, categories, and questions for those topics and categories. next step will be to do the actual seeding of data into the database.
