+++ package.json
@@ -17,7 +17,7 @@
     "seed": "node src/server/prisma/seed.js"
   },
   "dependencies": {
-    "@prisma/client": "^5.5.2",
+    "@prisma/client": "^5.11.0",
     "@reduxjs/toolkit": "^1.9.7",
     "bcrypt": "^5.1.1",
     "dotenv": "^16.3.1",
@@ -39,7 +39,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
     "nodemon": "^3.0.1",
-    "prisma": "^5.5.2",
+    "prisma": "^5.11.0",
     "vite": "^4.4.5"
   }
 }
